move observeList to an option of observe, make it use all the usual observer options, and spit up the observer code
